Working anon_alloc, write, exit in C userspace
The reason that the it was getting stuck when doing anon_allocs was due to the memory being too low and most likely running over some MMIO. To fix this I just moved the base of the process's virtual memory allocator up. Some hacky code has been implemented to set the TCBs. There are also some NULL checks on the filesystem operations.