process/thread: cache if userspace, thread->parent
The process structure now caches if it is initialized as a userspace
process or a kernel-space one. The thread's now use these values instead
of inferring it from page tables. Further, thread->parent is now set
when a thread is associated with a process, as it should've been from
the beginning.