I may keep around ProcessorDescriptor->process
It is quite convenient to use here. I have also made a small tweak in
the vm_map syscall. Instead of setting page_tables=NULL, they are
instead the user's page tables. However, this will bring up the problem
of a syscall handler trying to use such a mapping which will cause a
page fault and the fault handler mapping in a non-existent HHDM address
which will lead to further errors. Briefly, the generic page fault
handler needs to be extended to be able to resolve faults coming from
attempts to access userspace addresses too.