thread: per thread features
I have moved the needed processor features to be stored within the
thread instead of the process. This should enable the scheduler to more
quickly access this value and allow for a certain thread to have certain
features, but for another to not. I don't know how useful that will
actually be, but it is possible. For now though, all threads should have
the same features.